feat: run:function:start is now an alias to run:function:start:container
